+++ pcs_test/tier1/test_quorum.py
@@ -270,6 +270,8 @@ class DeviceAddTest(TestBase):
                 "model",
                 "net",
                 "host=127.0.0.1",
+                "tls=required",
+                "keep_active_partition_tie_breaker=on",
                 "algorithm=ffsplit",
                 "heuristics",
                 "mode=on",
@@ -287,6 +289,8 @@ class DeviceAddTest(TestBase):
                   Model: net
                     algorithm: ffsplit
                     host: 127.0.0.1
+                    keep_active_partition_tie_breaker: on
+                    tls: required
                   Heuristics:
                     exec_ls: test -f /tmp/test
                     mode: on
@@ -310,7 +314,7 @@ class DeviceAddTest(TestBase):
                 "algorithm=x c=d heuristics mode=bad e=f"
             ).split(),
             """\
-Error: invalid quorum device model option 'c', allowed options are: 'algorithm', 'connect_timeout', 'force_ip_version', 'host', 'port', 'tie_breaker', use --force to override
+Error: invalid quorum device model option 'c', allowed options are: 'algorithm', 'connect_timeout', 'force_ip_version', 'host', 'keep_active_partition_tie_breaker', 'port', 'tie_breaker', 'tls', use --force to override
 Error: 'x' is not a valid algorithm value, use 'ffsplit', 'lms', use --force to override
 Error: invalid quorum device option 'a', allowed options are: 'sync_timeout', 'timeout', use --force to override
 Error: '-1' is not a valid timeout value, use a positive integer, use --force to override
@@ -326,7 +330,7 @@ Error: Errors have occurred, therefore pcs is unable to continue
                 "algorithm=x c=d heuristics mode=bad e=f --force"
             ).split(),
             """\
-Warning: invalid quorum device model option 'c', allowed options are: 'algorithm', 'connect_timeout', 'force_ip_version', 'host', 'port', 'tie_breaker'
+Warning: invalid quorum device model option 'c', allowed options are: 'algorithm', 'connect_timeout', 'force_ip_version', 'host', 'keep_active_partition_tie_breaker', 'port', 'tie_breaker', 'tls'
 Warning: 'x' is not a valid algorithm value, use 'ffsplit', 'lms'
 Warning: invalid quorum device option 'a', allowed options are: 'sync_timeout', 'timeout'
 Warning: '-1' is not a valid timeout value, use a positive integer
@@ -639,8 +643,8 @@ class DeviceUpdateTest(TestBase):
             (
                 "Error: invalid quorum device model option 'c', allowed "
                 "options are: 'algorithm', 'connect_timeout', "
-                "'force_ip_version', 'host', 'port', 'tie_breaker', use "
-                "--force to override\n"
+                "'force_ip_version', 'host', 'keep_active_partition_tie_breaker',"
+                " 'port', 'tie_breaker', 'tls', use --force to override\n"
                 "Error: 'x' is not a valid port value, use a port number "
                 "(1..65535), use --force to override\n"
                 "Error: invalid quorum device option 'a', allowed options are: "
@@ -654,7 +658,7 @@ class DeviceUpdateTest(TestBase):
         self.assert_pcs_success(
             "quorum device update a=b timeout=-1 model port=x c=d --force".split(),
             """\
-Warning: invalid quorum device model option 'c', allowed options are: 'algorithm', 'connect_timeout', 'force_ip_version', 'host', 'port', 'tie_breaker'
+Warning: invalid quorum device model option 'c', allowed options are: 'algorithm', 'connect_timeout', 'force_ip_version', 'host', 'keep_active_partition_tie_breaker', 'port', 'tie_breaker', 'tls'
 Warning: 'x' is not a valid port value, use a port number (1..65535)
 Warning: invalid quorum device option 'a', allowed options are: 'sync_timeout', 'timeout'
 Warning: '-1' is not a valid timeout value, use a positive integer